Kinda fixed bug where player would spawn below level when loading level from MainMenu
I think this happens, because in the first ingame frame, the deltatime
is large enough for the player to move below the level in one frame,
but only when starting the game from the MainMenu.
I'm not quite sure why the deltatime is so high then.

The work-around solution:
Make all tiles `Loaded` by default. This prevents the player from moving
below the level in the first frame, and after that the
`EntityLoaderSystem` will unload the tiles appropriately again.