Hack together quick object spawner and EntityConfig SpawnAction
Although this seems to actually work,
I'm not sure if it's user-friendly enough.

The action `SpawnAction` spawns objects with the given
object data, which is the same data that comes from placing
an object in Tiled.
This makes the code quite re-usable, but configuration is
a bit unintuitive, because the Tiled generated data isn't meant
to be written by hand.

Issue #39