Add EventsRegisterData, add Delay event
There were (and still are) some problems with the `Interval` event:

1) Only one `Interval` event can be used per entity config,
   even if they have different milliseconds.
2) `Interval` events don't work in variants at all.
   More specifically, variants' `Interval` would only work if
   the base events has an `Interval` event defined; but
   because of problem 1), the base event's interval would always
   fire instead of any variant's interval.

Both of these problems are because the `HandleEventInterval` system
stores one `Timer` for every entity that has an `Interval` event.

The addition and usage of `EventsRegisterData` should prevent this,
because now each event can have some data associated (kinda).
And because this data is stored in the `EventsRegister`, it is cleared
when switching entities, which should prevent the problem with intervals not
working on variants.

The new `Delay` event uses the `EventsRegisterData` (and properly
handles multiple timers / multiple `Delay` events), so the above
problems aren't present with `Delay`. Just gotta fix up the
`Interval` event now.