Add entity config stack, add PushVariant and PopVariant actions to EntityAction
`EntityConfigRegister` now stores a entity config stack.
Any configs that are pushed or switched will change this stack accordingly.
Popping off will pop the last config off of the stack and apply the
second-to-last config on the stack (or the last, after popping).

Print warning message when attempting to switch or push
a variant that isn't defined. Also print message when
popping-off variant when there are no more configs on the stack.

Adjust `Discharger` enemy to use the new variant stack.

Closes #65