Add stupid ConsumableHealth custom entity
Entity that dies on contact with the player (or anything with
the "ConsumableHealthConsumer" collision tag).
The player gains health when it collides with the consumable
(with the tag "ConsumableHealth").

Pretty bad implementation, because both the consumable and the player
need to have logic, because they both have to do something.
But it works.

Issue #49