Add inherits field to entity config
Add inheritance to entity config.
Add abstract entity config settings.
Abstract entity configs are defined in `resources/entities/abstract/`.
Entity configs can inherit an abstract entity config with the `inherits` field.

Inheritance extends an entity config with abstract entity configs.
This happens once when loading the config.
So when loading an entity config, it is merged with all inherited
abstract configs specified in the `inherits` field.
The final inheritance-chain-merged config is then used ingame.
Inheritance is never used again during runtime, in contrast to variants,
which can be merged with the root entity config during runtime.

Inheritance is applied recursively, so an inherited abstract config can
have its own inheritance chain, that is also merged together.

Variants can also have inheritance chains.
Each variant's inheritance chain is applied once when loading its root
config, regardless of if the variant is gonna be active or not.
So basically, each variant is extended with its inherited abstract
configs, like how it works for the root config.

When merging entity config variants, each variant is merged with
existing variants now, instead of simply extending the variants HashMap.
This means, an abstract config can have a variant, and any config that
inherits this abstract config can define the same variant, which is then
merged with the abstract's variant of the same name.
Note, that events still need to be unique. So you overwrite an event if
you use the exact same event type in your config as is in the abstract config.

Issue #75