Change player entity config, update deathframe
Add "Control" variant to player entity config.
This variant is supposed to set all components and events,
that should be used in-game, when the user is in control of the player.
It should set stuff like gravity, friction, animations, anything that
can be re-applied anytime.
It should NOT have components like health, shooter, because that would
reset their health to full when re-applied. Those components should
be set in the root entity config's components field, to ensure that
they are only applied once when the player is spawned.

Player crouching now works better, because when player stops crouching
all components from the "Control" variant are re-applied.
Also animation is updated when switching crouched state,
because of deathframe update.

Issue #86