Refactor entity config components
Entity config `components` field is no longer a struct
with fields for every available entity config component.
Instead, `components` is now an array of `EntityConfigComponent`,
which is an enum with variants for every available entity config
component. Each variant takes an option of its respective component;
if you list the variant in the components array, but give it a `None`
value, then the component is _removed_, regardless of if it existed on
the entity or not.
This means, we can now _remove components_ using the `components` field.

Example new `components` syntax:
```ron
(
    // Notice that this is an array now ("[]" instead of "()").
    //         vvv
    components: [

    // Component name
    //  vvvvvvv
        Gravity(Some(   // <-
            y: -1000.0, // <- Component value (as option, `Some`)
        )),             // <-

        // You can omit the `Some` part.
        //    vvv
        Health((
            health: 10,
            max_health: 10,
        )),

        // Use `None` as a component value to _remove_
        // the component, in case it was present before.
        DealsDamage(None),
    ],
)
```

For this I had to fix the components syntax for every entity config we
have. While doing that I also refactored enemy configs a bit:
Added abstract enemy configs `"EnemyPhysics"`, `"EnemyHealth"`, and
`"Enemy"` (which just includes `"EnemyPhysics"` and `"EnemyHealth"`, and
adds the default enemy collision/solid tags).
This makes every enemy config way smaller and should make future enemy
tweaking easier.
See their abstract configs for more detailed descriptions.

While doing this, I also removed a lot of enemy components, that
I thought were unnecessary or would need tweaking anyway.
So a lot of enemies' components' values are fucked even more now.
At least it should be nicer/easier to tweak them now.

The `Size` component can also no longer be given in the `components`
array. Size has to be given in its own, new `size` field.
This is because the `Size` component has to be inserted before other
components, because some components rely on the size being present when
inserting. And because the `components` field is now an array, we can't
easily find and insert the `Size` component first (at least not without
some performance impact (probably), that I'm not willing to try out).

Example size config:
```ron
(
     size: (
        w: 16.0,
        h: 16.0,
     ),
     components: [
        // ...
     ],
)
```

Issue #48
Issue #78